chore: remove github actions send related code and fiber v3 package
