+++ .env.example
@@ -3,12 +3,6 @@
 # ======================
 BOT_TOKEN=
 APP_ID=
-BOT_CHANNEL_ID=
-
-# ======================
-# Fiber Config
-# ======================
-PRODUCTION_PORT=
 
 # ======================
 # Database Config
@@ -17,11 +11,6 @@ DB_NAME=
 DB_OWNER=
 DB_PASSWORD=
 DB_PORT=
-
-# ======================
-# Discord Group Config
-# ======================
-SELF_GROUP_ID=
 
 # ======================
 # VNDB Config
